+++ lib/screens/developer_settings_screen.dart
@@ -26,7 +26,6 @@ class DeveloperSettingsScreen extends HookConsumerWidget {
           child: WnSlate(
             header: WnSlateNavigationHeader(
               title: context.l10n.developerSettingsTitle,
-              type: WnSlateNavigationType.back,
               onNavigate: () => Routes.goBack(context),
             ),
             child: Padding(