+++ CHANGELOG.md
@@ -21,10 +21,11 @@ and this project adheres to Calendar Versioning (CalVer).
 ## Unreleased
 
 ### Added
-
+- Add bug report screen with NIP-44 encrypted anonymous reports [PR #478](https://github.com/marmot-protocol/whitenoise/pull/478)
 ### Changed
 - Deleted bubbles style [PR #471](https://github.com/marmot-protocol/whitenoise/pull/471)
 - Debug and relay screens improvements [PR #572](https://github.com/marmot-protocol/whitenoise/pull/502)
+- UI Polish: Slate & Settings Screen [PR #504](https://github.com/marmot-protocol/whitenoise/pull/504)
 - Show an error screen when bridge initialization fails instead of hanging on splash [PR #477](https://github.com/marmot-protocol/whitenoise/pull/477)
 
 ### Deprecated